requirements.txt
beautifulsoup4
requests
pdfminer.six
langchain
langchain-community
pypdf
python-dotenv
flask
